refactored the validators file to be a package to improve readability
