+++ src/validators/builtin_type_validator.py
@@ -1,0 +1,57 @@
+from collections import deque
+from collections import namedtuple
+
+from src.types import Data
+from src.types import RuleType
+from src.types import SchemaTypes
+from .base_validator import Validator
+
+_SchemaTypeDecoder = namedtuple("SchemaTypeDecoder", ["type", "friendly_name"])
+
+
+class BuildInTypeValidator(Validator):
+    """Validator to handle the build in types. e.g `int`, `list` & `str`"""
+
+    def __init__(self, violations: deque) -> None:
+        """BuildInTypeValidator init
+
+        Args:
+            violations (deque): Contains violations that have been detected
+            whilst processing the data
+        """
+        super().__init__(violations)
+        self._built_in_lookups = {
+            SchemaTypes.INT: _SchemaTypeDecoder(int, "int"),
+            SchemaTypes.STR: _SchemaTypeDecoder(str, "str"),
+            SchemaTypes.FLOAT: _SchemaTypeDecoder(float, "float"),
+            SchemaTypes.LIST: _SchemaTypeDecoder(list, "list"),
+            SchemaTypes.MAP: _SchemaTypeDecoder(dict, "map"),
+            SchemaTypes.BOOL: _SchemaTypeDecoder(bool, "bool"),
+        }
+
+    def validate(self, key: str, data: Data, parent: str, rtype: RuleType,
+                 is_required: bool = False) -> None:
+        """Validate the data against the core base types. If the data
+        does not match the expected type, then a `TypeViolation` is added
+        to the list of violations
+
+        Args:
+            key              (str): The key to the data
+            data            (Data): The data to validate
+            parent           (str): The parent key of the data
+            rtype       (RuleType): The type assigned to the rule
+            is_required     (bool): Is the rule required
+        """
+        buildin_type = self._built_in_lookups.get(rtype.type)
+        is_not_build_in_type = (buildin_type is None)
+
+        if is_not_build_in_type:
+            super().validate(key, data, parent, rtype, is_required)
+            return
+
+        if not isinstance(data, buildin_type.type):
+            message = f'{key} should be of type {buildin_type.friendly_name}'
+            self._add_type_violation(key, parent, message)
+            return
+
+        super().validate(key, data, parent, rtype, is_required)
