+++ src/validators/regex_validator.py
@@ -1,0 +1,37 @@
+from src.types import Data
+from src.types import RuleType
+from src.types import SchemaTypes
+from src.violations import RegexTypeViolation
+from .base_validator import Validator
+
+
+class RegexValidator(Validator):
+    """Validator to handle the regex type in a schema"""
+
+    def validate(self, key: str, data: Data, parent: str, rtype: RuleType,
+                 is_required: bool = False) -> None:
+        """Validate a regex rule type against some string data. If the data
+        is a string type, then a `TypeViolation` is added to the violation list.
+        If the string does not match the regex rule, then a `RegexTypeViolation`
+        is added to the violation list.
+
+        Args:
+            key              (str): The key to the data
+            data            (Data): The data to validate
+            parent           (str): The parent key of the data
+            rtype       (RuleType): The type assigned to the rule
+            is_required     (bool): Is the rule required
+        """
+        is_regex_type = (rtype.type == SchemaTypes.REGEX)
+        if not is_regex_type:
+            super().validate(key, data, parent, rtype, is_required)
+            return
+
+        if not isinstance(data, str):
+            self._add_type_violation(key, parent, f'{key} should be of type str')
+            return
+
+        if not rtype.regex.search(data):
+            violation = RegexTypeViolation(key, parent, data, rtype.regex)
+            self._violations.append(violation)
+            return
